Silence integer size conversion warnings in JavaHL under Win64 by casting the values explicitly.  We assume that argument counts and property sizes are all well below the 2G limit.
* subversion/bindings/javahl/native/CopySources.cpp
  (CopySources::array): cast size_t -> int
* subversion/bindings/javahl/native/CreateJ.cpp
  (CreateJ::Checksum, 
   CreateJ::PropertyMap): ditto
* subversion/bindings/javahl/native/StringArray.cpp
  (StringArray::array): ditto
* subversion/bindings/javahl/native/SVNClient.cpp
  (SVNClient::logMessages,
   SVNClient::merge,
   SVNClient::propertyGet,
   SVNClient::revProperty): same here ...
* subversion/bindings/javahl/native/Targets.cpp
  (Targets::array): ... and here

git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1405517 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/bindings/javahl/native/SVNClient.cpp
@@ -195,7 +195,7 @@ void SVNClient::logMessages(const char *path, Revision &pegRevision,
     SVN_JNI_ERR(target.error_occured(), );
 
     apr_array_header_t *ranges =
-        apr_array_make(subPool.getPool(), logRanges.size(),
+        apr_array_make(subPool.getPool(), (int)logRanges.size(),
                        sizeof(svn_opt_revision_range_t *));
 
     std::vector<RevisionRange>::const_iterator it;
@@ -644,7 +644,7 @@ void SVNClient::merge(const char *path, Revision &pegRevision,
         return;
 
     apr_array_header_t *ranges =
-      apr_array_make(subPool.getPool(), rangesToMerge.size(),
+      apr_array_make(subPool.getPool(), (int)rangesToMerge.size(),
                      sizeof(const svn_opt_revision_range_t *));
 
     std::vector<RevisionRange>::const_iterator it;
@@ -854,7 +854,7 @@ jbyteArray SVNClient::propertyGet(const char *path, const char *name,
         return NULL;
 
     return JNIUtil::makeJByteArray((const signed char *)propval->data,
-                                   propval->len);
+                                   (int)propval->len);
 }
 
 void SVNClient::properties(const char *path, Revision &revision,
@@ -1158,7 +1158,7 @@ jbyteArray SVNClient::revProperty(const char *path,
         return NULL;
 
     return JNIUtil::makeJByteArray((const signed char *)propval->data,
-                                   propval->len);
+                                   (int)propval->len);
 }
 void SVNClient::relocate(const char *from, const char *to, const char *path,
                          bool ignoreExternals)
